+++ xknx/remote_value/remote_value_sensor.py
@@ -10,7 +10,7 @@ from abc import abstractmethod
 from typing import TYPE_CHECKING, TypeVar, Union
 
 from xknx.dpt import DPTArray, DPTBase, DPTBinary, DPTNumeric
-from xknx.exceptions import ConversionError
+from xknx.exceptions import ConversionError, CouldNotParseTelegram
 
 from .remote_value import AsyncCallbackType, GroupAddressesType, RemoteValue
 
@@ -55,14 +55,14 @@ class _RemoteValueGeneric(RemoteValue[DPTArray, ValueType]):
             after_update_cb=after_update_cb,
         )
 
-    def payload_valid(self, payload: DPTArray | DPTBinary | None) -> DPTArray | None:
+    def payload_valid(self, payload: DPTArray | DPTBinary | None) -> DPTArray:
         """Test if telegram payload may be parsed."""
-        return (
-            payload
-            if isinstance(payload, DPTArray)
+        if (
+            isinstance(payload, DPTArray)
             and len(payload.value) == self.dpt_class.payload_length
-            else None
-        )
+        ):
+            return payload
+        raise CouldNotParseTelegram("Payload invalid", payload=str(payload))
 
     def to_knx(self, value: ValueType) -> DPTArray:
         """Convert value to payload."""