+++ xknx/remote_value/remote_value_string.py
@@ -6,6 +6,7 @@ DPT 16.000
 from __future__ import annotations
 
 from xknx.dpt import DPTArray, DPTBinary, DPTString
+from xknx.exceptions import CouldNotParseTelegram
 
 from .remote_value import RemoteValue
 
@@ -13,14 +14,14 @@ from .remote_value import RemoteValue
 class RemoteValueString(RemoteValue[DPTArray, str]):
     """Abstraction for remote value of KNX 16.000 (DPT_String_ASCII)."""
 
-    def payload_valid(self, payload: DPTArray | DPTBinary | None) -> DPTArray | None:
+    def payload_valid(self, payload: DPTArray | DPTBinary | None) -> DPTArray:
         """Test if telegram payload may be parsed."""
-        return (
-            payload
-            if isinstance(payload, DPTArray)
+        if (
+            isinstance(payload, DPTArray)
             and len(payload.value) == DPTString.payload_length
-            else None
-        )
+        ):
+            return payload
+        raise CouldNotParseTelegram("Payload invalid", payload=str(payload))
 
     def to_knx(self, value: str) -> DPTArray:
         """Convert value to payload."""